requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Django==4.1.5
django-crispy-forms==1.14.0
django-filter==22.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
Markdown==3.4.1
pycodestyle==2.10.0
python-dotenv==0.21.0
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
